Removed all implementations from state machine. Added double values support for calculations.

+++ calculator/src/main/java/com/EkkerArtem/stream/calculator/StreamCalculator.java
@@ -8,5 +8,5 @@ public interface StreamCalculator {
     /**
      * @return the calculated answer.
      */
-    Integer calculate();
+    double calculate();
 }
